requirements.txt
annotated-types==0.5.0
anyio==4.0.0
certifi==2023.7.22
click==8.1.7
distlib==0.3.7
fastapi==0.100.1
filelock==3.12.4
flake8==6.1.0
h11==0.14.0
httptools==0.6.0
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
packaging==23.1
pipenv==2023.9.8
platformdirs==3.10.0
pluggy==1.3.0
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
pytest==7.4.2
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.5
watchfiles==0.20.0
websockets==11.0.3
pymysql==1.1.0
